Ethernet with W5500 (#34)
* Adding the enet.c and enet.h to run the GrblHAL with an Ethernet cable using an W5500 module using telnet. The enet_start creates
a DHCP server and atributtes a given IP that can be changed to the computer connected to it. There are macros where you can
define your specific SPI pins, the ones there are the ones I was using. The driver.c was written by Terje and
it changes the way the HAL sends its interrupts system to fit with the DHCP server one.